+++ source-code/cljs/x/app_components/content.cljs
@@ -14,10 +14,10 @@
 ;; ----------------------------------------------------------------------------
 
 (ns x.app-components.content
-    (:require [app-fruits.reagent           :refer [component?]]
-              [mid-fruits.candy             :refer [param return]]
+    (:require [mid-fruits.candy             :refer [param return]]
               [mid-fruits.hiccup            :refer [hiccup?]]
               [mid-fruits.string            :as string]
+              [reagent.api                  :refer [component?]]
               [x.app-components.transmitter :rename {component transmitter}]
               [x.app-core.api               :as a :refer [r]]
               [x.app-dictionary.api         :as dictionary]))