+++ source-code/cljs/x/app_components/stated.cljs
@@ -14,9 +14,9 @@
 ;; ----------------------------------------------------------------------------
 
 (ns x.app-components.stated
-    (:require [app-fruits.reagent           :as reagent]
-              [mid-fruits.candy             :refer [param return]]
+    (:require [mid-fruits.candy             :refer [param return]]
               [mid-fruits.map               :refer [dissoc-in]]
+              [reagent.api                  :as reagent]
               [x.app-components.engine      :as engine]
               [x.app-components.subscriber  :rename {component subscriber}]
               [x.app-components.transmitter :rename {component transmitter}]