+++ source-code/cljs/x/boot_loader/side_effects.cljs
@@ -3,10 +3,10 @@
 ;; ----------------------------------------------------------------------------
 
 (ns x.boot-loader.side-effects
-    (:require [app-fruits.reagent :as reagent]
-              [dom.api            :as dom]
-              [x.app-core.api     :as a]
-              [x.app-ui.api       :as ui]))
+    (:require [dom.api        :as dom]
+              [reagent.api    :as reagent]
+              [x.app-core.api :as a]
+              [x.app-ui.api   :as ui]))